Remove the special @xxx imports and move them to lib/xxx
